Removed stdint.h dependency and added logging.
Not sure if it's worse or better without stdint.h types, oh well.
Logging is trying to catch a crashing bug that sometimes happens
during manipulator EXTEND with grabbed atom.
Bug has been difficult to reproduce, but my guess is that it has
to do with the the build_compound() process.